Added const version of TreeNode::get_children
TreeNode::get_children now creates a vector of const shared_ptr to ensure that, if desired, none of the children can be altered. TreeInternal::get_children now calls this function fixing an earlier shortcoming that only caused it to return a const vector instead of a vector of const shared_ptr.