TreeNode::get_children returns results depth first
Prior to this commit TreeNode::get_children returned the results by first giving the left child and its siblings. It then continued by appending the results from calls to TreeNode::get_children on the childeren. This is now changed to a depth-first search.

This is achieved by searching for the child on which the function is called. To prevent searching from start the last position is stored and search is started there.