Resolved build errors and delete and parent tests pass
Since some functions were removed the examples did not work anymore causing build errors. The examples have all be escaped with \dontrun{}.

The tests for parent and delete did not pass at first as a result of the migration to C++ backend. The tests have been adjusted and bugs resolved leading to all tests to pass successfully.