Extended tests sets for clone, getChildrenData and getSiblingsData
clone tests were not yet in repository.

getChildrenData now tests for named as well as non named arguments. 

getSiblingsData tests for warnings of deprecated methods as well as absolute getters.

expectations of  test_delete has been adjusted to reflect a more appropriate exception message thrown by GeneralTreeInternal.

GeneralTreeInternal now throws an exception that a key could not be found instead of id. This is more consistent with the nomenclature used throughout the package.