[cpp] [r] [t-r] [t-cpp] Improved node getters
Previously a separate function existed in R and C++ for each combination of key, data and leafs, children, sibling, branch. This has now been considerably simplified. A single C++ function exists for each type of node we want to retrieve. Internally and in the object a single R function exists to retrieve the information. 

All tests have been adjusted.